+++ src/main/java/ru/mikaev/sapr/domain/Rod.java
@@ -23,15 +23,15 @@ public class Rod {
     @GeneratedValue
     private Long id;
 
-    private int l;
+    private double l;
 
-    private int a;
+    private double a;
 
-    private int e;
+    private double e;
 
-    private int sigma;
+    private double sigma;
 
-    private int load;
+    private double load;
 
     @OneToOne(
             fetch = FetchType.EAGER,
